Update Client.h
There was a bug where WSARecv was attempting to receive 1024 byte chunks and the BYTE buffer was only 1024 bytes.
The issue is that in the OnRecv function, the byte at index dwNumberOfBytesTransfered was being set to null terminator.
If 1024 bytes were received, this would attempt to set the byte at index 1024 to '\0', but the array size is currently 1024 so the max element that can be accessed is at index 1023. To fix this, increase the initial buffer size by 1 byte.